Create object index and object mapid accessor functions
git-svn-id: svn://svn.code.sf.net/p/libgarmin/svn@95 1ee56e69-fe3a-0410-901e-914b4450df7a

+++ libgarmin/dev/src/libgarmin_priv.h
@@ -77,7 +77,7 @@ struct gimg {
 
 struct gpoint {
 	struct gar_subdiv *subdiv;
-	u_int32_t n;
+	u_int16_t n;
 	u_int8_t type;
 	u_int8_t subtype;
 	u_int32_t lbloffset;
@@ -91,7 +91,7 @@ struct gpoint {
 
 struct gpoly {
 	struct gar_subdiv *subdiv;
-	u_int32_t n;
+	u_int16_t n;
 	u_int8_t type;
 	u_int32_t lbloffset;
 	struct gcoord c;
